requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
flask>=3.0.0
google-generativeai>=0.8.0
